coord: rename catalog sql module to storage
The "sql" module name conflicted badly with the "sql" crate.
